+++ src/Tests.ToolKit/Data/FIleSystem/FileSystemRepositorySpecs/FileSystemRepositoryTests.cs
@@ -4,6 +4,7 @@ using FatCat.Fakes;
 using FatCat.Toolkit;
 using FatCat.Toolkit.Data.FileSystem;
 using FatCat.Toolkit.Json;
+using FatCat.Toolkit.Tools;
 
 namespace Tests.FatCat.Toolkit.Data.FIleSystem.FileSystemRepositorySpecs;
