+++ src/ToolKit/Communication/WebCaller.cs
@@ -1,6 +1,7 @@
 using System.Net;
 using FatCat.Toolkit.Console;
 using FatCat.Toolkit.Data;
+using FatCat.Toolkit.Logging;
 using FatCat.Toolkit.Web;
 using Flurl;
 using Flurl.Http;
