+++ src/ToolKit/ToolkitSettings.cs
